Fix unit tests - begin consolidation of type-checking functions in jxg-types.ts
